+++ CodeRed/DirectX12/DirectX12PipelineState/DirectX12BlendState.cpp
@@ -6,23 +6,26 @@ using namespace CodeRed::DirectX12;
 
 CodeRed::DirectX12BlendState::DirectX12BlendState(
 	const std::shared_ptr<GpuLogicalDevice> &device,
-	const BlendProperty& blend_property) :
-	GpuBlendState(device, blend_property)
+	const std::vector<BlendProperty>& blend_properties) :
+	GpuBlendState(device, blend_properties)
 {
-	const auto& property = mBlendProperties[0];
-
 	mBlendState.AlphaToCoverageEnable = false;
 	mBlendState.IndependentBlendEnable = true;
-	mBlendState.RenderTarget[0].BlendEnable = property.Enable;
-	mBlendState.RenderTarget[0].LogicOpEnable = false;
-	mBlendState.RenderTarget[0].SrcBlend = enumConvert(property.Source);
-	mBlendState.RenderTarget[0].DestBlend = enumConvert(property.Destination);
-	mBlendState.RenderTarget[0].BlendOp = enumConvert(property.ColorOperator);
-	mBlendState.RenderTarget[0].SrcBlendAlpha = enumConvert(property.SourceAlpha);
-	mBlendState.RenderTarget[0].DestBlendAlpha = enumConvert(property.DestinationAlpha);
-	mBlendState.RenderTarget[0].BlendOpAlpha = enumConvert(property.AlphaOperator);
-	mBlendState.RenderTarget[0].LogicOp = D3D12_LOGIC_OP_NOOP;
-	mBlendState.RenderTarget[0].RenderTargetWriteMask = enumConvert(property.ColorMask);
+
+	for (size_t index = 0; index < mBlendProperties.size(); index++) {
+		const auto& property = mBlendProperties[index];
+		
+		mBlendState.RenderTarget[index].BlendEnable = property.Enable;
+		mBlendState.RenderTarget[index].LogicOpEnable = false;
+		mBlendState.RenderTarget[index].SrcBlend = enumConvert(property.Source);
+		mBlendState.RenderTarget[index].DestBlend = enumConvert(property.Destination);
+		mBlendState.RenderTarget[index].BlendOp = enumConvert(property.ColorOperator);
+		mBlendState.RenderTarget[index].SrcBlendAlpha = enumConvert(property.SourceAlpha);
+		mBlendState.RenderTarget[index].DestBlendAlpha = enumConvert(property.DestinationAlpha);
+		mBlendState.RenderTarget[index].BlendOpAlpha = enumConvert(property.AlphaOperator);
+		mBlendState.RenderTarget[index].LogicOp = D3D12_LOGIC_OP_NOOP;
+		mBlendState.RenderTarget[index].RenderTargetWriteMask = enumConvert(property.ColorMask);
+	}
 }
 
 #endif