+++ CodeRed/Vulkan/VulkanPipelineState/VulkanBlendState.cpp
@@ -6,22 +6,24 @@ using namespace CodeRed::Vulkan;
 
 CodeRed::VulkanBlendState::VulkanBlendState(
 	const std::shared_ptr<GpuLogicalDevice>& device,
-	const BlendProperty& blend_property) :
-	GpuBlendState(device, blend_property),
+	const std::vector<BlendProperty>& blend_properties) :
+	GpuBlendState(device, blend_properties),
 	mAttachments(mBlendProperties.size())
 {	
-	const auto& property = mBlendProperties[0];
-	
-	mAttachments[0]
-		.setColorWriteMask(enumConvert(property.ColorMask))
-		.setBlendEnable(property.Enable)
-		.setAlphaBlendOp(enumConvert(property.AlphaOperator))
-		.setColorBlendOp(enumConvert(property.ColorOperator))
-		.setSrcColorBlendFactor(enumConvert(property.Source))
-		.setDstColorBlendFactor(enumConvert(property.Destination))
-		.setSrcAlphaBlendFactor(enumConvert(property.SourceAlpha))
-		.setDstAlphaBlendFactor(enumConvert(property.DestinationAlpha));
+	for (size_t index = 0; index < mBlendProperties.size(); index++) {
+		const auto& property = mBlendProperties[index];
 
+		mAttachments[index]
+			.setColorWriteMask(enumConvert(property.ColorMask))
+			.setBlendEnable(property.Enable)
+			.setAlphaBlendOp(enumConvert(property.AlphaOperator))
+			.setColorBlendOp(enumConvert(property.ColorOperator))
+			.setSrcColorBlendFactor(enumConvert(property.Source))
+			.setDstColorBlendFactor(enumConvert(property.Destination))
+			.setSrcAlphaBlendFactor(enumConvert(property.SourceAlpha))
+			.setDstAlphaBlendFactor(enumConvert(property.DestinationAlpha));
+	}
+	
 	mState
 		.setPNext(nullptr)
 		.setFlags(vk::PipelineColorBlendStateCreateFlags(0))